Update astrometry.net to 0.67+ (partially futurized version)
Has updates to the build system so that it can build on systems with python3,
but does not have any py3 updates to the python code itself.

Removed previous patch files since they fail to apply.

To better reconstruction the version of astrometry.net that was used
(since it isn't an official release), the tarball is named with the git
hash of the upstream master. Full hash for reference:
123ff3e168b86d1ba08fc90e3239d3068d747c24